refactor(Sprint6): ItemsPage를 styled-components로 교체, kebab-case 이름 변경
