--- vite-project/src/pages/ItemsPage.css
@@ -1,5 +1,0 @@
-.ItemsPage {
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-}
